added routes for comparing drivers teams

+++ main.py
@@ -14,6 +14,17 @@ firebase_request_adapter = requests.Request()
 app.mount("/static", StaticFiles(directory="static"), name="static")
 templates = Jinja2Templates(directory="templates")
 
+def is_user_logged_in(request:Request):
+    id_token = request.cookies.get("token")
+    if not id_token:
+        return False
+    try:
+        user_token = google.oauth2.id_token.verify_firebase_token(id_token, firebase_request_adapter)
+        if user_token:
+            return True
+    except Exception as e:
+        return False
+
 @app.get("/", response_class=HTMLResponse)
 async def root(request: Request):
 
@@ -27,6 +38,7 @@ def retrive_drivers(request:Request,
                   attribute: Optional[str] = None, 
                   operator: Optional[str] = None, 
                   value: Optional[str] = None):
+    user_logged_in = is_user_logged_in(request)
     if attribute and operator and value:
         query = firestore_db.collection('drivers')
         try:
@@ -39,8 +51,10 @@ def retrive_drivers(request:Request,
         query_info = f"{attribute} {operator} {value}"
         return templates.TemplateResponse(
             "display_drivers.html",
-            {"drivers":drivers,
+            {   "request":request,
+                "drivers":drivers,
              "query_info":query_info,
+             "user_logged_in":user_logged_in,
             "request": request}
         )
     else :
@@ -49,7 +63,8 @@ def retrive_drivers(request:Request,
         return templates.TemplateResponse(
             "display_drivers.html",
             {"drivers":drivers,
-            "request": request}
+            "request": request,
+             "user_logged_in":user_logged_in,}
         )
 
 
@@ -92,6 +107,8 @@ def register_drivers(
     fastest_laps: str = Form(...),
     drive_team: str = Form(...)
 ):
+    if not is_user_logged_in(request):
+        return RedirectResponse(url="/", status_code=303)
     # Convert string inputs to integers
     try:
         age = int(age)
@@ -167,6 +184,8 @@ def register_drivers(
     fastest_laps: str = Form(...),
     drive_team: str = Form(...)
     ):
+    if not is_user_logged_in(request):
+        return RedirectResponse(url="/", status_code=303)
     try:
         age = int(age)
         pole_positions = int(pole_positions)
@@ -215,11 +234,83 @@ def register_drivers(
 
 @app.delete("/drivers/{driver_id}",response_class=HTMLResponse)
 def delete_driver(request:Request,driver_id:str):
+    if not is_user_logged_in(request):
+        return RedirectResponse(url="/", status_code=303)
     ref = firestore_db.collection("drivers").document(driver_id)
     if not ref.get().exists:
         raise Exception("Driver not found")
     ref.delete()
     return RedirectResponse(url="/drivers", status_code=303)
+
+@app.get("/compare_drivers")
+async def compare_drivers_page(request: Request):
+    try:
+       
+        drivers_ref = firestore_db.collection('drivers')
+        drivers = []
+        for doc in drivers_ref.stream():
+            team_data = doc.to_dict()
+            team_data['id'] = doc.id
+            drivers.append(team_data)
+        
+       
+        return templates.TemplateResponse("compare_drivers.html", {
+            "request": request,
+            "drivers": drivers
+        })
+    except Exception as e:
+        
+        return HTMLResponse(content="something went wrong", status_code=404)
+    
+@app.post("/compare_drivers")
+async def compare_drivers(request: Request, driver1: str = Form(...), driver2: str = Form(...)):
+    try:
+        
+        driver1_doc = firestore_db.collection('drivers').document(driver1).get()
+        driver2_doc = firestore_db.collection('teams').document(driver2).get()
+        
+        
+        if not driver1_doc.exists or not driver2_doc.exists:
+            raise Exception("Team not found")
+        
+        
+        driver1_data = driver1_doc.to_dict()
+        driver1_data['id'] = driver1_doc.id
+        
+        driver2_data = driver2_doc.to_dict()
+        driver2_data['id'] = driver2_doc.id
+
+        
+        if driver1_data['id'] == driver2_data['id']:
+            raise Exception("comparison cannot be done between two same entities")
+        
+       
+        comparison = {}
+        comparison['year_founded'] = 'driver1' if driver1_data['year_founded'] < driver2_data['year_founded'] else 'driver2'
+        comparison['pole_positions'] = 'driver1' if driver1_data['pole_positions'] > driver2_data['pole_positions'] else 'driver2'
+        comparison['race_wins'] = 'driver1' if driver1_data['race_wins'] > driver2_data['race_wins'] else 'driver2'
+        comparison['constructor_titles'] = 'driver1' if driver1_data['constructor_titles'] > driver2_data['constructor_titles'] else 'driver2'
+        comparison['prev_finish_position'] = 'driver1' if driver1_data['prev_finish_position'] < driver2_data['prev_finish_position'] else 'driver2'
+
+        
+        result = {
+            'driver1': driver1_data,
+            'driver2': driver2_data,
+            'comparison': comparison
+        }
+        
+       
+        return templates.TemplateResponse("compare_teams_result.html", {
+            "request": request,
+            "driver1": result['driver1'],
+            "team2": result['team2'],
+            "comparison": result['comparison']
+        })
+    except Exception as e:
+        
+        return HTMLResponse(content=str(e), status_code=404)
+
+
 
 
 
@@ -233,6 +324,7 @@ def retrive_teams(request:Request,
                   attribute: Optional[str] = None, 
                   operator: Optional[str] = None, 
                   value: Optional[str] = None):
+    user_logged_in = is_user_logged_in(request)
     if attribute and operator and value:
         query = firestore_db.collection('teams')
         try:
@@ -247,6 +339,7 @@ def retrive_teams(request:Request,
             "display_teams.html",
             {"teams":teams,
              "query_info":query_info,
+             "user_logged_in":user_logged_in,
             "request": request}
         )
     else :
@@ -255,6 +348,7 @@ def retrive_teams(request:Request,
         return templates.TemplateResponse(
             "display_teams.html",
             {"teams":teams,
+             "user_logged_in":user_logged_in,
             "request": request}
         )
 
@@ -279,6 +373,8 @@ def update_team(request:Request,
                 race_wins: str = Form(...),
                 constructor_titles: str = Form(...),
                 prev_finish_position: str = Form(...)):
+    if not is_user_logged_in(request):
+        return RedirectResponse(url="/", status_code=303)
     try:
         year_founded = int(year_founded)
         pole_positions = int(pole_positions)
@@ -333,6 +429,8 @@ def register_team(
     constructor_titles: str = Form(...),
     prev_finish_position: str = Form(...)
 ):
+    if not is_user_logged_in(request):
+        return RedirectResponse(url="/", status_code=303)
     
     try:
         year_founded = int(year_founded)
@@ -383,8 +481,80 @@ def query_teams(request:Request,
 
 @app.delete("/teams/{team_id}",response_class=HTMLResponse)
 def delete_team(request:Request,team_id:str):
+    if not is_user_logged_in(request):
+        return RedirectResponse(url="/", status_code=303)
+
     ref = firestore_db.collection("teams").document(team_id)
     if not ref.get().exists:
         raise Exception("Team not found")
     ref.delete()
     return RedirectResponse(url="/teams", status_code=303)
+
+
+@app.get("/compare_teams")
+async def compare_teams_page(request: Request):
+    try:
+       
+        teams_ref = firestore_db.collection('teams')
+        teams = []
+        for doc in teams_ref.stream():
+            team_data = doc.to_dict()
+            team_data['id'] = doc.id
+            teams.append(team_data)
+        
+       
+        return templates.TemplateResponse("compare_teams.html", {
+            "request": request,
+            "teams": teams
+        })
+    except Exception as e:
+        
+        return HTMLResponse(content="something went wrong", status_code=404)
+    
+@app.post("/compare_teams")
+async def compare_teams(request: Request, team1: str = Form(...), team2: str = Form(...)):
+    try:
+        
+        team1_doc = firestore_db.collection('teams').document(team1).get()
+        team2_doc = firestore_db.collection('teams').document(team2).get()
+        
+        
+        if not team1_doc.exists or not team2_doc.exists:
+            raise Exception("Team not found")
+        
+        
+        team1_data = team1_doc.to_dict()
+        team1_data['id'] = team1_doc.id
+        
+        team2_data = team2_doc.to_dict()
+        team2_data['id'] = team2_doc.id
+
+        
+        if team1_data['id'] == team2_data['id']:
+            raise Exception("comparison cannot be done between two same entities")
+        
+       
+        comparison = {}
+        comparison['year_founded'] = 'team1' if team1_data['year_founded'] < team2_data['year_founded'] else 'team2'
+        comparison['pole_positions'] = 'team1' if team1_data['pole_positions'] > team2_data['pole_positions'] else 'team2'
+        comparison['race_wins'] = 'team1' if team1_data['race_wins'] > team2_data['race_wins'] else 'team2'
+        comparison['constructor_titles'] = 'team1' if team1_data['constructor_titles'] > team2_data['constructor_titles'] else 'team2'
+        comparison['prev_finish_position'] = 'team1' if team1_data['prev_finish_position'] < team2_data['prev_finish_position'] else 'team2'
+
+        
+        result = {
+            'team1': team1_data,
+            'team2': team2_data,
+            'comparison': comparison
+        }
+        
+       
+        return templates.TemplateResponse("compare_teams_result.html", {
+            "request": request,
+            "team1": result['team1'],
+            "team2": result['team2'],
+            "comparison": result['comparison']
+        })
+    except Exception as e:
+        
+        return HTMLResponse(content=str(e), status_code=404)
